requirements.txt
mysql-connector-python==9.0.0
scikit-learn==1.6.1
numpy==2.2.2
flask==3.1.0
flask-cors==5.0.0
python-dotenv